+++ Tests/Maths/Graph/DynamicSubgraphTest.cpp
@@ -31,7 +31,7 @@ namespace sequoia::testing
     }
 
     {
-      graph_test_helper<complex<float>, complex<double>, test_subgraph>  helper{*this};
+      graph_test_helper<complex<double>, complex<double>, test_subgraph>  helper{*this};
       helper.run_tests();
     }
   }
